Reorganized files and folders to group the Subclasses of a class better. Created items and enemies folder to save the different subclasses of those classes.
