+++ retroBattle.h
@@ -13,7 +13,7 @@
 // TODO: Verweisen Sie hier auf zusätzliche Header, die Ihr Programm erfordert.
 #include "headers/entity.h"
 #include "headers/Player.h"
-#include "headers/Enemy.h"
+#include "headers/enemies/Enemy.h"
 #include "headers/BattleManager.h"
 #include "headers/Vec2D.h"
 #include "headers/EntityManager.h"
